enhance: collapse device management routes into singular admin/devices route
