+++ ui/user/src/routes/admin/device-clients/+page.ts
@@ -1,43 +1,6 @@
-import { handleRouteError } from '$lib/errors';
-import {
-	AdminService,
-	type DeviceClientSortKey,
-	type DeviceClientFleetSummaryResponse,
-	type OrgUser,
-	UserService
-} from '$lib/services';
-import { profile } from '$lib/stores';
 import type { PageLoad } from './$types';
-import { DEFAULT_SORT_BY, DEFAULT_SORT_ORDER, sortFields } from './constants';
+import { redirect } from '@sveltejs/kit';
 
-function getSortBy(property: string | null): DeviceClientSortKey {
-	const key = property ?? '';
-	return Object.hasOwn(sortFields, key) ? sortFields[key] : DEFAULT_SORT_BY;
-}
-
-function getSortOrder(order: string | null): 'asc' | 'desc' {
-	return order === 'desc' ? 'desc' : 'asc';
-}
-
-export const load: PageLoad = async ({ fetch, url }) => {
-	const limit = parseInt(url.searchParams.get('pageSize') ?? '50', 10) || 50;
-	const offset = parseInt(url.searchParams.get('offset') ?? '0', 10) || 0;
-	const name = url.searchParams.get('name') ?? '';
-	const sortProperty = url.searchParams.get('sort');
-	const hasValidSortProperty = sortProperty != null && Object.hasOwn(sortFields, sortProperty);
-	const sortBy = getSortBy(sortProperty);
-	const sortOrder = hasValidSortProperty
-		? getSortOrder(url.searchParams.get('sortDirection'))
-		: DEFAULT_SORT_ORDER;
-	let clients: DeviceClientFleetSummaryResponse;
-	let users: OrgUser[];
-	try {
-		[clients, users] = await Promise.all([
-			AdminService.listDeviceClients({ limit, offset, name, sortBy, sortOrder }, { fetch }),
-			UserService.listUsers({ fetch })
-		]);
-		return { clients, users };
-	} catch (err) {
-		handleRouteError(err, '/admin/device-clients', profile.current);
-	}
+export const load: PageLoad = async () => {
+	throw redirect(301, '/admin/device-overview?view=device-clients');
 };
