--- ui/user/src/routes/admin/device-overview/constants.ts
@@ -1,4 +1,0 @@
-// DEFAULT_WINDOW_MS is the rolling time range applied to the device
-// overview when the URL doesn't pin start/end. Mirrors the backend
-// dashboardWindowDefault in pkg/api/handlers/devicescans.go.
-export const DEFAULT_WINDOW_MS = 60 * 24 * 60 * 60 * 1000;
