+++ ui/user/src/routes/admin/devices/+page.ts
