+++ ui/user/src/routes/admin/devices/constants.ts
@@ -1,5 +1,10 @@
 import type { DeviceClientSortKey } from '$lib/services/admin/types';
 
+// DEFAULT_WINDOW_MS is the rolling time range applied to the device
+// overview when the URL doesn't pin start/end. Mirrors the backend
+// dashboardWindowDefault in pkg/api/handlers/devicescans.go.
+export const DEFAULT_WINDOW_MS = 60 * 24 * 60 * 60 * 1000;
+
 export const defaultSort = {
 	property: 'name',
 	order: 'asc'
